Add files via upload
This code is a simple Python function block that outputs the string "Hello, GitHub!". To use this code on GitHub, you need to create a new repository, add a .py file, and copy this code into the file.

Please note that this is only a minimal example, and the actual code on GitHub will be more complex and include many other elements such as descriptions, comments, guides, etc.